fixed renaming of entities.
when an entity is renamed, the index in the corresponding entity container (of type renderer::EntityVector or renderer::EntityMap) gets out-of-date.  we might want to fix that in the future; in the meantime, entities are always deleted then recreated using appleseed::studio::ProjectBuilder::replace_entity() , even when just their name has changed.  consequently appleseed::studio::ProjectBuilder::edit_entity() has been removed.

another change was the renaming of the name and model parameters in the dictionary used to keep track of entity parameters during creation/edition, in order to reduce the risk of parameter name clashes: they are now called "__name" and "__model" instead of "name" and "model".