some code simplifications and optimizations in progressive renderer.
moved the vector of samples from renderer::SampleGeneratorJob to renderer::SampleGenerator to avoid reallocating it each time a new job is created.

the sample generators are now created each time rendering begins (and destroyed when rendering ends) instead of being created once when the renderer::ProgressiveFrameRenderer is constructed. with this change, renderer::SampleGenerator::reset() is no longer necessary and has been removed.