changed the definition of normalized device coordinates throughout appleseed.renderer.
until now, normalized device coordinates (NDC) were defined in the domain [-0.5,0.5) x (-0.5,0.5] with (-0.5,-0.5) located in the bottom-left corner in the image and (0.5,0.5) located in the top-right corner. not only was this definition non-symmetrical, but the fact that it considered points with y=0.5 to be part of the domain made it sometimes necessary to clamp y coordinates to prevent them from overflowing.

now, they are simply defined in [0,1)^2 with (0,0) located in the top-left corner of the image and (1,1) located in the bottom-right corner.