fixed a major flaw in the way materials are assigned to object instances.
until now, object instances were assigned an array of material *indices* which would obviously become invalid as soon as materials are reordered or deleted. in particular, when a material is edited, it is first deleted then recreated at the end of the material array, immediately invalidating all material assignments in the current assembly.

with this commit, object instances are now assigned an array of material names. when rendering starts, material names are translated to material pointers for fast lookup.

other changes in this commit:

  - renderer::Camera::on_frame_begin() is no longer passed an intersector; cameras that need one (e.g. for autofocus) will need to create it themselves.
  - scenes and assemblies are now responsible for calling on_frame_begin() and on_frame_end() on their child entities.
  - renderer::Project::get_trace_context() is now const.