fixed computation of average luminance, added computation of rms deviation against reference image, cleaned up renderer::Frame.
it is now possible to set the path to a reference image in the progressive renderer.  when a reference image is set, a plot of the rms deviation between the current image and the reference image is recorded, and the current rms deviation and average luminance error is reported during rendering.

during progressive rendering, the average frame luminance is now computed on the final frame (after clamping and conversion to the output color space) and no longer on the linear RGB frame.  on the one hand it makes absolute luminance values somewhat meaningless if the output color space is not linear RGB, on the other hand it enables easy and accurate comparisons of the average luminance with that of a (necessarily clamped and color space-transformed) reference image.

additional refactorings:

  - added renderer::Frame::image() to access the underlying foundation::Image object.
  - removed redundant tile(), properties(), compute_average_luminance() and compute_rms_deviation() methods from renderer::Frame.
  - moved the statistics printing code from renderer::AccumulationFramebuffer to the progressive renderer.