removed a dangerous trap in foundation::PoolAllocator and cleaned up the code.
the trap:  the page size of a pool allocator was specified in bytes, leading to deadly issues if misused (e.g. if the page size is not a multiple of the item size or if the page size is too small to accommodate a single item).  with this commit, the page size is now specified in number of items rather than bytes.

the cleanups:
- disabled the assignment operator
- removed the default value for the page size
- added a unit test for rebinding from a pool allocator with the void value type to a pool allocator of a different value type, added comments
- simplified some unit tests