reimplemented render layers for lights, EDFs and environment EDFs.
also added support for render layers to the distribution ray tracer lighting engine.

the contribution of a light, an EDF or an environment EDF can be stored in its own framebuffer by setting a "render_layer" attribute on the entity:

  <edf name="white_light_edf" model="diffuse_edf">
      <parameter name="exitance" value="white_light_exitance" />
      <parameter name="render_layer" value="light_group_1_color" />
  </edf>

many entities can contribute to the same render layer, but a given entity can contribute to at most one render layer.