reverted a2ac6285756bdc3073ddea410f9fb4e274c56d06.
Reverted a2ac6285756bdc3073ddea410f9fb4e274c56d06: delete widgets immediately in appleseed::studio::clear_layout(). The reason is that we often have signals in flight when we clear layouts (for instance when we rebuild the entity editor form when changing the model of a BSDF), so deleting widgets immediately is incorrect and leads to crashes.