Added ability to create DisneyMaterials
Editing works as it should. After creating a Disney Material it's own
editor will be used.
Created MaterialCreationDelayedAction for scheduled creation of
the materials as well. Even if most of the code seems duplicated,
it is necessary since the architecture of the interface is different
from what needs to be added.

Signed-off-by: Marius Avram <marius.avram1309@gmail.com>