Parameter can have value between tags
Until now there was only the possiblity to add a value for a parameter
using the value attribute. This had the disadvantage of not allowing
the preservation of some special characters such as newlines.

The solution to this problem is to enclose the value between a pair
of tags of the same type: e.g. <parameter name="metallic">15</parameter>.

This is particularly useful in the case of disney expressions.

Signed-off-by: Marius Avram <marius.avram1309@gmail.com>