Importing and exporting disney materials
They are exported using individual files. Currently the settings.xsd
format is used and some extra parameters are added to save the name
and the model of the materials.

Contains also a refactoring of show_message_box() from expressioneditorwindow
which now is required in multiple places. It has been moved to tools.cpp
and renamed to show_warning_message_box().

Signed-off-by: Marius Avram <marius.avram1309@gmail.com>