implemented light targets and photon targets.
a light target is a volume of interest.  light targets act as hints to the lights' sampling method.  when no light target is defined, the entire scene is considered of interest.

photon targets are light targets used by the SPPM engine to shoot photons toward volumes of interests (such as a window in a wall).

a photon target can be defined by adding a "photon_target" parameter with value "true" to an object instance:

    <object_instance name="window_inst" object="window">
        <parameter name="photon_target" value="true" />
        ...
    </object_instance>