Fix the build against Clang 3.4.1 and above
Add missing namespace specifiers (boost::) when name resolution can be
ambiguous (it is already done for shared_ptr's).  Modern C++ standard
libraries typically would implement mutex and thread types, which will
conflict with the ones from Boost.  This is required for the upcoming
FreeBSD support.

This change has no impact on GCC (verified for versions 4.7.4, 4.8.5,
and 4.9-20150819) and early versions of Clang (3.3.0).