Fix brightness spread mismatch
Maya uses as 2D noise, value noise, not gradient noises which are
provided by OSL (Perlin, simples). Furthermore it uses a 1D random
noise as part of the brightness spread feature, and this was causing a
considerable mistmatch. Approximate as best as we can this node's
appearance. Adding value noise might improve this in the future.