Fix area light shader, checker range
The arealight shader was not using the maya convention (camel case) and
as result, was not found, this was fixed and the chance taken to extend
the shader with the expected outputs for the area light node in Maya.
The checkerboard node was producing values outside [0,1] resulting in
valid samples being discarded.