Add UNDEFINED_UVFILTER definition
When trying to test uvFilterSize (float2) OSL fails since it cannot use
isconnected() on an array, nor on the elements of an array, but it would
be handy to use the UV filtersize instead of calling filterwidth()
repeatedly whenever we need this quantity. We can test for an invalid
quantity and do the right thing however.