Add turbulence noise, REQUIRE_VALUE_NOISE
When using value noise, a large rng table is needed, which slows down
shader compilation considerably. Since the only shader that requires
value noise for now is the mountain node, set a definition before the
includes so that the table is only set with the data when the definition
is set.