Add maya fBm, turbulence
The turbulence fractal used in Maya noise looks too regular, but it
matches better if the shading point is perturbed with another noise.
This gives it a more boiling/cloud like appearance. As for fractional
Brownian motion, some extra numbers were added to try and bring the
result of the fBm into something that resembles better the Maya output.