Add back time global primvar to space/time noise
The time, dtime global primvars were not set and filled with garbage, so
the noise node space-time mode was ignoring them and using only the maya
UI input value. This was fixed upstream with est77/appleseed-maya@b82048c1fad4cea54296674ad523fb215d354b98 so it can be re-enabled and DEBUG statements removed.