Fix Blinn roughness parameter
The as_glossy_blinn closure squares the roughness parameter internally
so one needs to pass the square root here, but the closure is defined in
the extensions header, before any of the standard OSL functions. Deal
with this by extending the microfacet() closure for the "blinn" MDF,
only for the isotropic BRDF case.