Moved layeredShader to its own branch
This requires a custom exporter and attributes to deal with both color
and closure types as inputs. To make matters worse, these are compound
data types with arrays as elements, and array element connections are
still unsupported in OSL. With these requirements and potential changes
it's better to move this to its own branch.