Add chromatically adapted XYZ<>RGB matrices
The previous matrices were assuming the white points set in the
respective technical recommendations, but when transforming spaces we
need to account for the different white points. For now support only
D60, D65, DCI, which are the most relevant for our purposes and should
address the immediate needs for ACES 2065-1 AP0, ACEScg AP1, Rec.2020,
Rec.709/sRGB, DCI-P3, D60-P3. The respective luminance coefficients are
also dependent on the chromatically adapted matrices (Y value).