Rename temp dicrectory for gathering packed project files
Name of temp directory was the same as name of unpacked directory.
It led to clashing of this two directories.
If we opened packed project and tried to save it again it failed with error
because it tried to create "project.unpacked" directory which was already existed.
New name doesn't have this issue.