Fix CIE Lab,Luv,LCh, remapping, glass, see notes
CIE L*a*b* and CIE L*u*v* return values with L in [0,100] range, and a,b
and u,v in [-100,100] range. Color samples however were discarded for
being negative, and it's not really easy to adjust this in a normal
workflow, so CIELAB, CIELUV are mapped to [0,1] overall.
The CIE LCh function was also not correct, and is now fixed, and
remapped as well.
Finally, the glass shader was querying ray_depth, before proceeding with
tracing the main BSDF. This was not working, so it was removed, and at
the same time, we add the check for "shadow" and "transparency" rays to
bypass the shader completely.