Complete color pipeline overhaul
The color pipeline is now configurable to use RGB colors or full spectra. The color pipeline mode ("spectrum mode") is stored in a thread-local static variable. All DynamicSpectrum<> objects now always have the same size within a thread. We no longer upgrade RGB colors to spectra or downgrade spectra to RGB colors. DynamicSpectrum<>::resize() is gone.

Spectra intent are now only use when converting between RGB colors and spectra, and are no longer stored. Consequently, the size of a 31-channel DynamicSpectrum<> is now exactly 128 bytes (31 floats + 1 float of padding).

The result of shading a pixel sample is now always in linear RGB and is represented by a foundation::Color4f object. The classes ShadingFragment, ShadingFragmentStack and SpectrumStack are gone.