OSL tweaks for standard, texture, triplanar
Address single and multiple channel IO in asTexture. This fixes #1903.
Added generic relative luminance Y, taking as parameter the working
space specification, assuming as white point, the white point used in
the working space standard, and assuming the input color already had an
EOTF applied and was converted to the same RGBW primaries as the working
space.