Add constructors for the derived AABB class
With GCC 8, the keyword using is not working as expected and we need to
implement all the constructors required in the derived class. Moreover,
the keyword using is not available by default with VS2012.
The constructor using the base class as a parameter is required to use
all static members using or returning base class objects.