Improve adaptive tile renderer settings and behavior
- Change adaptiveness to min_samples. This allow to show a fixed number
of uniform samples. Like this we don't need to change adaptiveness when
we change max_samples.
- Allow to set max_samples to 0. This will stop the render only when
pixels have converged. Though you should use a max_samples when caustics
are enabled or when difficul feature may never converge or provide high
variance.
- When computing the noise threshold of a block, use the maximum instead
of the sum. It makes the renderer much more predictable and
user-friendly. Values are in a more coherent range.