Tweaks to ASC CDL help, parameters
The parameters for the RGB and white primaries is not *colorspace* but
RGB primaries and white, so name it accordingly.
The parameters for primaries and clamping should not be mapeable and
visible in the node graph editor, so hide them accordingly.
Parameter names have *in_* or *out_* prefix, but the clamp parameter had
an *as_* prefix, which is incorrect. The convention is this prefix being
used for appleseed generic shader names (not DCC specific).