Fixes for invertColor, see detail log below
The maya attribute names cannot have spaces between them, and the
convention specifies explicitly camelCase for the long attribute names.
The short attribute names are meant to be 3 character long maximum.
The shader and node name as invertColor, but abbreviations were used for
documentation and icon, so use the exact same name for consistency.
The color color doesn't need to be a mapper, a ordinary popup with the
color models suffices.
The popup and the checkboxes to invert the channels were allowed to be
mapped. The missing attributes connectable, keyable, hidden were added,
all set to 0 except hidden set to 1. Blender input socket also set to 0
as well as gafferNoduleLayoutVisible since these parameters are not to
be exposed in the node graph view.
Avoid unnecessary transformc() when the input color model is RGB since
transformc() is expensive.
Set explicit origin space in the transformc() calls, and add DEBUG
clause to catch any other model (in case we add more later).
Rename channel parameters, avoiding the 0,1,2 numbering and explicitly
using X, Y, Z components of the vector type (color).