Implement render stamp scaling
Add new scale factor control numeric to the stamp postprocessing
metadata.

Use the appleseed_seeds_64 float array as source data for the icon.
Scale the original 64 x 61 pixel image data to fit into the scaled area
using OpenImageIO's ImageBuf and resize function. Blit the scaled data
onto the frame. Also scale the text before drawing.

The mitchell filter is used for the resize function with the filter size
scaling according to the user scale factor as this gave best visual
results.

The default height of the destination region of interest is set to 14 to
best match the previous icon size.