Use volume to account for crop window
From my testing, the crop window is initialized to the image boundaries even if one hasn't been explicitly set by the user, which is nice.

Also, attempt to avoid integer overflow since the default value is the numeric limit. After a bit of research, we could add a helper to detect integer overflow here and assign a max value, but otherwise it might be just unreasonable for this to ever be an issue. For instance, a 1mil x 1mil pixel image still could have over 10mil spp.