Add Pixel Error AOV, improve rendering error analysis
See https://www.ci.i.u-tokyo.ac.jp/~hachisuka/mbdpt.pdf for an elaborate reference example.

Generally useful for understanding the behavior of sampling or light transport optimizations. For instance, we would expect there to be lower error near hard to reach specular spots for some bidirectional methods (SPPM, BDPT...). An obvious next step would be to enable comparing two renders side-by-side normalized against the highest error of a single render (or allow an upper bound param), but this might be better suited as a Python script.

Directly relevant to the discussion in #2465 regarding the behavior of computing tile noise in sRGB space, where we might expect to see less variance in darker areas and higher variance in lighter areas for sRGB.

Usage:

    <frame name="beauty">
        ...
        <parameter name="reference_image" value="ref.exr" />
        <aovs>
            <aov model="pixel_error_aov" />
      </aovs>
    </frame>