Implement texture controlled image plane sampling
Add new pixel renderer class TextureControlledPixelRenderer created by
new class TextureControlledPixelRendererFactory.
TextureControlledPixelRenderer determines the number of iterations
of the sample loop by querying a texture for that pixel position.
Textures can be of any resolution, and are resized to exactly match the
frame dimensions for a 1:1 frame/texture correspondence. Resizing will
also change the aspect ratio of the texture if it doesn't match the
frame, possibly squishing the texture in one dimension.

Since the texture driving the sampling should be black and white anyway,
only the red channel of the texture is used to retrieve values.

Add new setting controls to the GUI to set min/max samples, the force
anti-aliasing parameter and a file path for the
TextureControlledPixelRenderer. The decorrelate option is ignored since
it is set to be removed in a pending PR (#2467).