Change the pause/resume text based on rendering state
When the render is rendering, the Pause/Resume menu item
can lead to confusion as it is not clear which state the renderer
is in based on a checkbox. Does the checkbox mean that the renderer
is running, or does it mean that it is paused? This change makes it
clear which state the renderer is in. If the renderer is running
the menu item will say 'Pause Rendering', likewise if the renderer
is paused the menu item will say 'Resume Rendering'. When the rendering
is canceled or finished, the menu item will revert back to
'Pause/Resume Rendering'.