Add texture-controlled pixel renderer
Add new `TextureControlledPixelRenderer` pixel renderer class.

The number of iterations of the sample loop is determined by querying a texture for that pixel position. Textures can be of any resolution and are resized to exactly match the frame dimensions for a 1:1 frame/texture correspondence. Resizing will also change the aspect ratio of the texture if it doesn't match the frame.

Since the texture driving the sampling should be black and white anyway, only the red channel of the texture is used to retrieve values.