Add path guiding UI settings
Add new GuidedPathTracerPanel to the render settings window. The
render settings of the uniform sampler do not work ideally with path
guiding yet. In the rendercomponents the samples setting is for now
passed to the GPTPassCallback so it can end rendering. The
GPTPassCallback will quit rendering when the number of passes rendered
times the number of samples per pass surpasses the total number of
samples set by the user. The samples per pass setting in the guided path
tracer panel is passed to the UniformPixelRendererFactory. The
passes parameter for the GenericFrameRendererFactory are overwritten
with a high number (10000000) of passes in the rendercomponents, so the
PassManagerFunc's pass loop can go on until stopped by the
GPTPassCallback. This is not working at this point because the passes
parameters are already accessed before the rendercomponents leading to
the remaining render time estimation being completely off and samples
from passes beyond the UI passes setting not contributing to the image.
This can only be prevented by setting an arbitrarily high number of
passes in the UI.
These issues will need a later overhaul but with these caveats in mind
the guided path tracer can be controlled from the UI.