Add pass callback logic
GPTPassCallback keeps track of passes and iterations and initiates a new
iteration with twice the number of passes if the current iteration
finished. On a new iteration a variance projection based on the current
frame's accumulated samples is calculated. The renderer either
starts a new SD-tree refinement iteration with double the pass count if
the projection indicates that the variance is continuously decreasing
between iterations, or renders the remaining sample budget if the
variance is projected to not be improved by further SD-tree refinement.

Add VarianceTrackingShadingResultFrameBuffer together with new factory.
The new framebuffer inherits from ShadingResultFrameBuffer and adds a
new channel to keep track of the summed squared samples which are used
for the variance calculation.

Add virtual get_total_channel_count() method to
ishadingresultframebufferfactory interface and derived classes to allow
dynamically determining the number of channels based on the factory.
These should now be used instead of the corresponding static method in
ShadingREsultFrameBuffer.

To make the guided path tracer fully checkpoint compatible it is not
enough to only store the shading buffer contens, there will also have to
be a mechanism added for dumping and reloading the SD-tree.