Fix variance estimation
Fix variance estimation in VarianceTrackingShadingResultFramebuffer.
Previously I divided the sample sum as well as the sum of squared
samples by the weight sum (which is the number of samples for that
pixel), but then divided by the number of samples again for the
estimate. The division is now only done for the estimate.

Add method variance_image() and variance_to_tile() to the factory and
framebuffer to write the variance estimate for each pixel to an image.
This is needed for the optional sample combination between path guiding
iterations.