Various path guiding changes
Allow unlimited path-guided bounces. Guided bounces are treated as
Diffuse or Glossy internally but do not contribute to the respective
bounce counters. Remove ScatteringMode::has_diffuse_or_glossy added in
previous commit as it is not needed with this handling.

Add comments and references to SD-tree implementation.

Pass bsdf sampling fraction parameter to path visitor directly instead
of setting and getting it in the guided path.