D-Tree samples return ScatteringMode
On rebalancing the D-Tree after each training iteration an appropriate
ScatteringMode will be assigned to the D-Tree.
D-Tree sample directions will be treated as Glossy if a significant
fraction of the trees energy is stored in a small subset of directions
over the sphere and as Diffuse otherwise. D-Tree samples are then
treated according to their ScatteringMode by the rest of the system
without.

Trees that did not collect radiance during the previous iteration are
now reset back to default untrained state.