Split input section of BSDFSample into BSDF::LocalGeometry
Important:

With this commit, cos(incoming direction, shading normal) is now computed with the unmodified (unperturbed) shading normal when using OSL, whereas before this commit it was computed with the modified one.

This produces better looking bump mapping, and it also matches the way bump mapping is computed by the built-in shading system.