Reduce Critical Section When Developing Accumulation Buffer to Frame
When using interactive rendering the rendering is restarted often when updating the camera transformation.
A lot of time is spend developing the AccumulatorTile to the Frame at higher active frame levels. This is a serial bottleneck when rendering a frame.
By copying the AccumulatorTile in the critical section and developing outside,
more time can be spend in the renders threads, resulting in a less noisy overal image when interacting.
The cost is double the memory for the LocalSampleAccumulationBuffer.