Alternative: post process the frame just after rendering
This is an alternative solution to the bug "False Colors not being applied to all tiles".

The problem is the following:
1. We emit a signal for the end of the render **before** applying post-process.
    The signal is emitted using `renderer_controller.on_rendering_success()` in `MasterRenderer::do_render()`.
2. Because of that, appleseed.studio apply False Colors before post-process.
3. False Colors gets applied on a copy of the frame, to keep the original render.
4. So if there is a post-process + False Colors, False Colors get replaced by the post-process.

One solution is to call `on_rendering_success` after applygin post-process, which order-correctly the effects:
1. Post-process is applied first
2. Then False Colors is applied on the post-processed frame
3. It works !

The problem with that solution, is that we don't call `on_rendering_success` and `on_rendering_abort` from `do_render` anymore, but `on_rendering_begin()` still is. Trigerring the signals from differents places in the code make it less easy to maintain and understand.

I think we should keep the structure in `do_render()` to make sure interactive rendering + final rendering always work the same way.

Another solution, the one in this commit, is to apply the post-process right after render. Then, no matter when you emit the signal, the frame will always be ready and will never get modified after the signal is being emitted.