correct project modified when crop region changes
The project will now be marked as modified if a new crop region is defined or if the crop region is cleared.

The slot slot_project_mofified() of mainwindow.cpp was used as a function since the function creating the crop Area/ render region was already in the same class.